moved Font settings view from SettingsActivity to FontSettingsController
